Adds initial port of ziggyrama stg2 to 2010 STi
This is as direct a port as was possible with the existing ECU
definitions for the 2010 STi. Differences include:

* per-gear ignition timing compensation tables were not ported from the
2011 map because the 2010 ECU definition doesn't include those tables
* per-gear requested torque tables were not ported, which ziggyrama says
were updated to avoid overboosting in higher gears. These tables don't
appear in either RomRaider or ECU Edit.
* size of CL Fueling Target Compensation is different. Removed 1 row and
1 column to fit 2010 table.

This direct port overboosts significantly on my car with the Cobb intake
and Harman Motive downpipe installed (21-ish psi).